1.1.0

+++ modules/Task/Target.js
@@ -12,17 +12,34 @@ module.exports = exports = {
 
     //同步目录结构。
     syncDirs(meta) {
-        let { source, target, } = meta;
+        let { source, target, console, stat, } = meta;
         let dir$name = {}; //同步后的 target 中对应于 source 的目录列表。
+        let timer = new Timer(console);
+
+        timer.start(`开始同步目录结构，共 ${colors.cyan(source.dirs.length)} 个>>`.bold);
 
         //先同步目录结构。
         source.dirs.forEach((dir) => {
             let name = Path.relative(source.dir, dir);
             let dir2 = Path.join(target.dir, name, '/');
 
-            Directory.create(dir2);
+            if (!fs.existsSync(dir2)) {
+                console.log(`  创建目录:`, dir2);
+
+                if (!meta.simulate) {
+                    Directory.create(dir2);
+                }
+              
+                stat.createDirs.push(dir2);
+            }
+
+           
             dir$name[dir2] = name;
+            
         });
+
+        timer.stop(`<< 目录结构同步完成，耗时{text}`.bold);
+
 
         return dir$name;
 
@@ -32,19 +49,9 @@ module.exports = exports = {
 
     //同步文件列表。
     syncFiles(meta) {
-        let { console, source, target, } = meta;
+        let { console, source, target, stat, } = meta;
         let { files, } = source;
         let file$md5 = {};
-
-        let stat = {
-            successCopys: [],
-            failCopys: [],
-            successRenames: [],
-            failRenames: [],
-            jumpFiles: [],
-        };
-
-
         let bar = new ProgressBar(files, console);
         let timer = new Timer(console);
 
@@ -65,7 +72,11 @@ module.exports = exports = {
             if (!files || !files.length) {
                 try {
                     bar.log(` `, `复制文件`.green, file.gray);
-                    File.copy(file, file2);
+
+                    if (!meta.simulate) {
+                        File.copy(file, file2);
+                    }
+                    
                     stat.successCopys.push(file);
                 }
                 catch (ex) {
@@ -96,7 +107,10 @@ module.exports = exports = {
                 bar.log(``, `重命名文件`.yellow, file1.yellow);
                 bar.log(`       `, `-->`, file2.underline);
 
-                fs.renameSync(file1, file2);
+                if (!meta.simulate) {
+                    fs.renameSync(file1, file2);
+                }
+
                 files.splice(0, 1); //删除它，表示已处理。
                 stat.successRenames.push(file1);
             }
@@ -109,38 +123,37 @@ module.exports = exports = {
 
         timer.stop(`<< 文件同步完成，耗时{text}`.bold);
 
-        return { file$md5, stat, };
+        return file$md5;
     },
 
 
     //清理。
     clear(meta, { dir$name, file$md5, }) {
-        let { console, target, } = meta;
-        let { dirs, files, } = FileSystem.scan({ console, ...target, });
+        let { console, target, patterns, stat,} = meta;
+        let { dirs, files, } = FileSystem.scan({ console, ...target, patterns, });
         let dirKeys = Object.keys(dir$name);    //需要保留的目录列表。
         let fileKeys = Object.keys(file$md5);   //需要保留的文件列表。
-
-        let stat = {
-            deleteDirs: [],
-            deleteFiles: [],
-        };
-
         let timer = new Timer(console);
 
 
         timer.start(`清理 target 目录 >>`.bold);
 
+        //模仿同步的，要无条件地执行。
+
         //性能优化，先判断长度。
         //实际的目录数多于要保留的目录数，则执行删除。
-        if (dirs.length > dirKeys.length) {
+        if (meta.simulate || dirs.length > dirKeys.length) {
             dirs.forEach((dir) => {
                 let name = dir$name[dir];
                 if (name) {
                     return;
                 }
 
-               
-                Directory.delete(dir);
+
+                if (!meta.simulate) {
+                    Directory.delete(dir);
+                }
+
                 console.log(`  `, `删除目录`.bgMagenta, dir.magenta);
                 stat.deleteDirs.push(dir);
             });
@@ -148,7 +161,7 @@ module.exports = exports = {
 
         //性能优化，先判断长度。
         //实际的文件数多于要保留的文件数，则执行删除。
-        if (files.length > fileKeys.length) {
+        if (meta.simulate || files.length > fileKeys.length) {
             files.forEach((file) => {
                 let md5 = file$md5[file];
 
@@ -156,7 +169,10 @@ module.exports = exports = {
                     return;
                 }
 
-                File.delete(file);
+                if (!meta.simulate) {
+                    File.delete(file);
+                }
+               
                 console.log(`  `, `删除文件`.bgMagenta, file.magenta);
                 stat.deleteFiles.push(file);
             });
@@ -164,7 +180,6 @@ module.exports = exports = {
 
         timer.stop(`<< 清理 target 目录完成，耗时{text}。`.bold);
 
-        return stat;
 
     },
